test: splitting the auth tests out into its own file
