+++ __tests__/__fixtures__/createOas.js
@@ -1,0 +1,20 @@
+module.exports = serverUrl => {
+  return function (method = 'get', path = '/', operation = {}) {
+    return {
+      openapi: '3.0.0',
+      info: {
+        title: 'OAS test',
+      },
+      servers: [
+        {
+          url: serverUrl,
+        },
+      ],
+      paths: {
+        [path]: {
+          [method]: operation,
+        },
+      },
+    };
+  };
+};
